refactor: delete static cast

+++ src/robocup_client/sender/node/sender_node.cpp
@@ -72,12 +72,9 @@ void SenderNode::publish_unit()
   auto gyro = robot_client->get_gyro();
   auto accelero = robot_client->get_accelero();
 
-  unit_msg.gyro = {static_cast<float>(gyro.value().x()), static_cast<float>(gyro.value().y()),
-    static_cast<float>(gyro.value().z())};
+  unit_msg.gyro = {gyro.value().x(), gyro.value().y(), gyro.value().z()};
 
-  unit_msg.accelero =
-  {static_cast<float>(accelero.value().x()), static_cast<float>(accelero.value().y()),
-    static_cast<float>(accelero.value().z())};
+  unit_msg.accelero = {accelero.value().x(), accelero.value().y(), accelero.value().z()};
 
   this->unit_publisher->publish(unit_msg);
 }
